refactor: Changes in generate_tt command
Added time and date arguments. The check on incorrect input has now more informaion. Deleted the Weekends check and set the time limits to 20:30 in different way. New argument to the command: date and time. Flow argument is now obligatory argument. Renamed few variables, thinking of renaming the functions and rework this module.